Add reference samples from RFC1808; Clarify empty reference resolution;
The question was whether an empty reference removes the base fragment or not.
RFC 1808 kept the original reference (explicit sample), RFC 2396 C.2. stated vaguely "(current document)" instead of a concrete value, and the original xproto code seemed conflicted.
One test case wanted to keep the fragment, yet a commented out test case referenced vaguely that RFC 3986 demands to remove the fragment with seemingly cited text.
That text, however, cannot be found. Yet the pseudocode in 5.2.2. clearly states that regardless of the conditions, the reference fragment is always taken.